Publish Engineer degrees to graduate collection in DSpace
Why these changes are being introduced:

In the SqsMessage model, Engineer degrees are incorrectly assigned
the handle for the DSpace undergrad theses collection.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-468

How this addresses that need:

This updates the collection handle for Engineer degrees to the
graduate theses collection.

Side effects of this change:

None.