Add report for thesis with ProQuest forms
Why these changes are being introduced:

Thesis processors need to be able to download ProQuest forms.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-481

How this addresses that need:

This creates a report that lists the theses with ProQuest forms.
Each row includes a clickable link of the ProQuest form for processors
to download.

Side effects of this change:

Adds a new route and modifies the ability model to account for the
new controller action.