Add report on theses created by students
Why these changes are being introduced:

There is a need to see which thesis records were created by students
rather than the registrar data import.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-416

How this addresses that need:

* Sets paper_trail's whodunnit feature in Application Controller,
which will track which user is responsible for each version of a
record.
* Creates a report that lists which theses' first version has a
whodunnit, filtered by term. This ensures we won't see theses
created by the registrar import job, as whodunnit is nil for those
records.
* Sets whodunnit to 'registrar' for all thesis and userrecords created
by the registrar import job. This is intended to create a clearer audit
trail for system-generated records (otherwise, they would all have
a nil value for whodunnit).

Side effects of this change:

* Strictly speaking, the report technically does not _only_ show
theses created by students. The `student?` method checks whether
a user has the basic role and not the admin flag, which should
cover nearly all thesis submitters. However, if a department admin
with the transfer submitter role is also a student and deposits
their thesis, that will not appear in this report.
* Whodunnit will be tracked for _all_ records, not just theses,
and for all versions. This prepares us for a possible future state
when additional audit trail reporting may be needed.
* The `modified_by` method is now in the User Helper, renamed as
`link_to_user` without the paper_trail context so it can be used
in other situations. It also now links to the admin edit page to
match the title links in reports.