Fix mailer bug in DSpace Publication Results Job
Why these changes are being introduced:

In its current state, the app does not send mail if the DSpace
Publication Results Job is run and the results include errors only.
This is because mail is sent when results[:processed] is positive,
and this condition will never be met if the job returns no successes.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-502

How this addresses that need:

This changes the condition for the DSpace Publication Results Job
to send mail. Mail will now be sent either if results[:total] is
positive, or if results[:errors] has any items. The results[:errors]
piece of this is necessary because the job can fail before it starts
polling, in which case results[:total] will not increment.

Side effects of this change:

I noticed some TODOs in the DSpace Publication Job around setting
the publication status to 'Publication error'. It seems we didn't
have that status in the app when we enabled that job. We do now,
so I updated the job and corresponding tests accordingly.