Clearly show absence of degrees and advisors in processing form
Why these changes are being introduced:

Theses require degrees and advisors to be valid for publication,
but it's not clear in the UI when a thesis is missing these fields
because the form hides them if they are empty.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-476

How this addresses that need:

This adds readonly values for the advisor and degree form elements
that inform the user if the thesis has no associated advisors or
degrees.

Side effects of this change:

* Refactored the degree field UI to a partial, as the additional logic
complicated it enough to warrant it.
* Added JavaScript to remove the readonly advisor element when an
advisor is added, to avoid inadvertently saving the "no advisor" message
as an advisor in the database.