Add report on theses by publication status
Why these changes are being introduced:

Processors need a report that clearly shows the current publication
status of the theses in the application.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-315

How this addresses that need:

This creates a report that lists theses' current publication statuses.
The report includes param filters to optionally refine the results by
term, publication status, or both. DataTables is included for additional
filtering if needed.

Side effects of this change:

Adds a CSS rule to create some spacing between the two selects.