Further describe DSS status errors
Why these changes are being introduced:

When the DSpace Publication Results Job sees a record with a result
type of error, it adds the exception message and the thesis ID to
the errors array. We've seen at least once instance where the result
type is error but the exception message is empty, which results in
an unhelpful message like this: " (thesis ID 42)".

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-510

How this addresses that need:

This adds a bit more context to the error message described above,
so can determine the following:

1. That DSS encountered an error while processing the record; and
2. What the publication status was updated to (hopefully 'error').

This also updates the field where exception messages appear in the
result message, as this recently changed in DSS.

Side effects of this change:

This will change (i.e., enrich) how errors are reported in the
publication results email.