Use UTF-8 character instead of code for copyright symbol
Why these changes are being introduced:

DSpace does not support UTF-8 codes or HTML entities, so the
only way to display a copyright symbol is to use the UTF-8
character itself.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-515

How this addresses that need:

This uses the UTF-8 character in the Dspace Metadata mdoel instead
of the UTF-8 code.

Side effects of this change:

None expected, but we'll need to determine whether this will publish
successfully. I've confirmed that DSpace test renders the character
when editing the record metadata in the DSpace UI, but I haven't
confirmed that the full publication workflow will run as expected.