Add dspace_handle to administrate thesis dashboard
Why these changes are being introduced:

Ordinarily dspace_handle will be automatically generated as part of
the publication workflow, but there are some cases in which a thesis
must be published manually, which would require the processor to input
the handle in administrate.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-519

How this addresses that need:

This makes dspace_handle a form attribute in the administrate thesis
dashboard.

Side effects of this change:

None.