Add text mapping for "No Creative Commons" license type
Why these changes are being introduced:

During publication workflow testing, stakeholders noted that "No
Creative Commons" looks odd in DSpace, and it would be preferable
for these license types to read as "In Copyright -- Educational Use
Permitted" instead.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-514

How this addresses that need:

This adds a method to the license model that returns the desired
phrase for "No Creative Commons" license types, and otherwise
returns the license type as stored in the database. We then use
that method in the Dspace Metadata model to ensure that the mapped
value appears in DSpace.

Side effects of this change:

This change complicates the testing approach we use for the Dspace
Metadata model, which effectively checks for all possible values of
dc.rights in the unserialized metadata against a given thesis'
copyright and license. Now that we are using the same value for two
different copyright/license states, the tests are a little more
ambiguous. This isn't a problem yet, but it's worth documenting here.