Adds a rake task for processing dss output queue
Why are these changes being introduced:

* We need a way to run the output processing job

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-501

How does this address that need:

* Creates a rake task to allow us to easily trigger the job to run on
  Heroku from our local CLI rather than having to create a console
  session in Heroku to run job

Document any side effects to this change:

If multiple apps are connected to the same output queue
(such as local dev and staging), running in one (such as local dev) will
only update the local instance but possibly be processing output for
different environemnts. In practice, this should not be a risk in
production as nobody should have configuration locally to connect to
production output queues so the risk is more associated with PR builds,
Staging and Development environemnts which seem low enough risk to have
the benefits outweigh this concern.