Add license URL to theses with no-CC license
Why these changes are being introduced:

In the DspaceMetadata class, we map thesis records with author
copyright and no CC license to a dc.rights value of “In Copyright -
Educational Use Permitted.” However, because the no-CC license
record has no URL associated, these thesis records show up in
DSpace with no dc.rights.uri field. We want the dc.rights.uri
field to match the statement, so we need to associate this rights
URL with no-cc license theses: 'In Copyright - Educational Use
Permitted | Rights Statements'.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-534

How this addresses that need:

This adds an evaluate_license_url method to the License model that
returns the desired URL if the corresponding license_type is equal
to "No Creative Commons License". This method is then invoked in
DspaceMetadata#copyright, in the condition branch for which the
author holds copyright and provides a license.

Side effects of this change:

Hopefully minimal, as this approach will limit the scope of the
effects. (The alternative of modifying the database directly would
affect _all_ records with the no-cc license, which might yield
unexpected results.)