Create structure for BagIt bags
Why these changes are being introduced:

We need a data structure that creates the foundation for preservation
bags.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-530

How this addresses that need:

This creates a Bag model that takes a thesis and generates generates
a bag declaration, manifest and bag name for it. It also includes
a `data` method that returns a hash of filepaths for the bag's data
directory, along with their corresponding blobs.

Side effects of this change:

* The base64_to_hex method has been moved to a new Checksums module,
as it is now used both in the publication results job and in this
new model. We may choose to move additional checksum logic to this
module if it is required in multiple places.
* Once ETD-529 is complete, we will need to build the metadata CSV
in this class and add it to the data hash. It's possible this can
all happen within the `data` method, depending on how that ticket
is implemented.
* While this PR meets the need of the ticket conceptually speaking,
additional refinements will likely be required as we determine how
we will create the bags and save them to S3.