Refactor Bag class
Why are these changes being introduced:

In the future we may need to know when a thesis is sent to preservation and
what files were included in its BagIt bag. A thesis might also be sent to
preservation more than once, so a record of each preservation submission
is needed.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-555
* https://mitlibraries.atlassian.net/browse/ETD-556

How does this address that need:

This refactors the Bag class into SubmissionInformationPackage, a model
that achieves similar functionality but with ActiveRecord backing it. When
SubmissionInformationPackage is created, it generates a bag name,
bag declaration, metadata CSV, and manifest, and saves them all to the db.

The SubmissionInformationPackage model belongs to the Thesis model, and
a Thesis has many SubmissionInformationPackages. This allows us to account
for the same thesis being sent to preservation multiple times.

Finally, to track provenance, paper_trail is enabled for this model and the
preserved_at and preservation_status properties are added.