Create and attach zipped bag to SIP
Why are these changes being introduced:

* delivery of a bag to s3 is how we are handling preservation

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-531
* https://mitlibraries.atlassian.net/browse/ETD-558

How does this address that need:

* Creates a temporary zip file on disk and streams objects for a thesis
  into that tempfile
* Attaches that tempfile via ActiveStorage to a newly created
  SubmissionInformationPackage

Document any side effects to this change:

* This uses Heroku's filesystem to store the tempfiles to create the zip
which is non-ideal but because this all happens in a single event it
should be safe to do this. The use of a tempfile ensures the file is
deleted when the handle is closed so if this occurred over multiple
requests it would already fail for non-Heroku reaons.

* This streams objects from S3 into Heroku and back as a zip file. This
will cost data transfer in both directions. A more cost effective
solution if this proves expensive would be to move the logic to create
the zip files to AWS and run it in the same region as the S3 buckets
which would remove the transfer costs.