Create batch jobs for preservation
Why these changes are being introduced:

We need a way to initiate SIP creation, bagging and streaming to
S3.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-553

How this addresses that need:

This creates two jobs:
1. PreservationSubmissionJob attempts to send a single thesis to
preservation and modifies the preservation_status and preserved_at
fields accordingly.
2. PreservationSubmissionPrepJob loops through an array of theses
to preserve.

Side effects of this change:

* preservation_status is now an enum with three options: unpreserved
(the default), preserved, and error.
* Adds `preservation.rake` task to send a single thesis to
preservation via Rails CLI.
* The readme now indicates that the publishing workflow automatically
kicks off the preservation workflow. This is not yet accurate, but
will be once ETD-560 is implemented.