Update location of metadata CSV in SIPs
Why these changes are being introduced:

The preservation package spec requires that metadata files be
stored as `data/metadata/metadata.csv`, but they are currently
stored as `data/metadata.csv.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-564

How this addresses that need:

This fixes the location of the metadata CSV in the
SubmissionInformationPackage model.

Side effects of this change:

None.