Guard against nil tempfiles in MARC export job
Why these changes are being introduced:

If a tempfile has already been closed and unlinked, the `ensure`
block in MarcExportJob#perform will throw an error.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-546

How this addresses that need:

This checks if the tempfile exists before attempting to close it.

Side effects of this change:

Adds rubyzip to gemfile. Rubyzip is already used by other gems, and
it can be called in local development, but bootsnap seems to require
it to be loaded explicitly.