Add holds by source report
Why these changes are being introduced:

We need to be able to see the holds for a given degree period,
filtered by hold source.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-462

How this addresses that need:

This adds a report of holds that allows filtering on thesis
grad date and hold source.

Side effects of this change:

* Report#extract_terms now checks for the type of items in the
collection, so as to accommodate non-thesis collections.
* Performance may be slow due to possible N+1 queries. We do
not have the test data to confirm this, but we will make changes
as needed depending on its behavior in production.

Co-Authored-By: Jeremy Prevost <jprevost@mit.edu>