Fix bug in summary report for authors not graduated
Why these changes are being introduced:

QA on #916 revealed a discrepancy between the number
of authors not graduated in the summary report and
the authors not graduated report. This is because we
were not deduping thesis records in
Report#data_authors_not_graduated.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-429

How this addresses that need:

In Report#data_authors_not_graduated, this calls `select` instead
of `map` to identify ungraduated authors, then calls `uniq` on the
result to avoid counting duplicate records.

Side effects of this change:

This went unnoticed until QA in part because we do not extensively
test the summary report. It's probably okay to test that feature
minimally since this sort of thing is likely to be caught by
stakeholders, but we should consider extending unit testing for
reporting if it happens again.