Dedup authors not graduated summary data
Why these changes are being introduced:

Theses with multiple files are being counted for each attached
file in the authors not graduated summary data.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-429

How this addresses that need:

This dedups theses to fix the count in the Report index.

Side effects of this change:

N/A.