Pass environment to Sentry
Why these changes are being introduced:

All Sentry alerts currently display the environment as 'production'.
We would like Sentry to accurately show the environment in which
the alert was captured.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-181

How this addresses that need:

This sets the Sentry environment to `SENTRY_ENV`, defaulting to
'unknown' if that variable isn't set.

Side effects of this change:

None.