Use shared Ruby CI and specify Heroku stack
Why these changes are being introduced:

We want to update our applications to use the Heroku 22
stack, which also requires us to use Ubuntu 22.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-180
https://mitlibraries.atlassian.net/browse/ENGX-182

How this addresses that need:

This updates the CI config to point to the shared workflow
and adds the Heroku stack version to app.json.

Side effects of this change:

To accommodate the shared CI workflow, the SimpleCov output
filename is now specified.