Disable export_concurrent in test_helper
Why these changes are being introduced:

We occasionally encounter a segfault during precompile, which is
caused by Sprockets' `export_concurrent` feature. This issue
occurred intermittently before using the shared Ruby CI workflow,
but it now seems to happen on every CI run.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-180

How this addresses that need:

This disables `export_concurrent` in the test helper.

Side effects of this change:

The test helper might not be the most logical place for this config
change. We tried doing it in test.rb, but that did not yield
consistent results.

co-authored-by: Adam Jazairi <adamjazairi@gmail.com>