Refactor administrate authorization to support gem update
Why these changes are being introduced:

The most recent version of administrate refactors authorization
in a way that has downstream effects. See [#1941](https://github.com/thoughtbot/administrate/pull/1941)
for more info.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-192

How this addresses that need:

This replaces our override of the `show_action?` method with
`authorized_action?` in the administrate application controller,
and it replaces our combined check of `valid_action?` and
`show_action?` with `accessible_action?` in the administrate views.

This also updates the ability so thesis processors can continue to
access the submitter dashboard. This is not currently needed
as we do not have any users with the thesis_processor role, but it
seemed wise to address it sooner than later.

Side effects of this change:

* Thesis processors can now see links to controller actions in the
submitter admin dashboard, as well as a link to the admin dashboard
itself from the main site nav.
* The admin submitter dashboard test will move to the admin/
subdirectory with the other individual dashboard tests. This change
will come in a subsequent commit to make the git history easier
to read.