Add ProQuest opt-in to thesis submission forms
Why these changes are being introduced:

We need to document permission from authors of graduate theses as to
whether we can send their theses to ProQuest or not.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-587

How this addresses that need:

This adds a proquest_allowed boolean field to the author model thesis and
adds it as a field to the new and edit thesis forms. As per the ticket
requirements, this field only shows in the edit form for graduate theses,
which includes theses that satisfy both graduate and undergradate
degree types. Null values are allowed in the data model, but the field is
validated in the form.

Side effects of this change:

* The new field has been added to the author administrate dashboard. This
was not specified in the success criteria, but it seemed like it would
be useful to processors.
* A version of the errorPlacement jQuery function used in the processing
form has been added to the submission forms. This is needed to place the
validation error for the radio button fields in a readable location.
* The new form element will need UXWS review before we deploy it to
production.