Initial set of reporting features for ProQuest opt-in status
Why these changes are being introduced:

Processors need to be able to review ProQuest opt-in status
information at two points in the processing workflow:

1. Prior to publication but after files have been attached,, so they
can review and resolve discrepancies in opt-in status for theses with
multiple authors.
2. After publication, so they can review summary data by
department, degree, and term.

During grooming, these use cases were conceived as separate reports,
but combining them into a single report seemed to be a more
usable design.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-591
* https://mitlibraries.atlassian.net/browse/ETD-599

How this addresses that need:

This adds a report that lists the ProQuest opt-in status for theses
with files attached. The report includes five filters:

* Graduation date (i.e., term)
* Department
* Degree
* Multiple authors (boolean)
* Published theses (boolean)

Additionally, the report displays cards listing summary data for each
status. If a status is included in the filtered results, then a card
will list the number of theses with that status that match all of
the conditions.

Side effects of this change:

* As this is a slight departure from what we discussed in grooming,
it may not be the optimal approach for stakeholders and may need
to be redesigned pending QA.
* The summary data cards do not hook into the existing card system
in the Report model. This particular use case was custom enough that
it seemed more efficient to start from scratch.
* Some existing fixtures have been altered, requiring unrelated
count-based tests to be modified.
* Assertions were added to some unrelated report controller tests
that didn't seem to test anything.
* In our reports, we use the extract_terms method to show only
the terms for which we have theses. I chose not to do this with
degrees and departments, as it may be useful for processors to see
all of them. If there is not an actual use case for this, then it
will be trivial to add similar extract methods for degrees and
departments.
* The multiple_authors scope introduces counter cache to the
application. This use case seems well suited to counter cahce, but
it's worth calling out here as something we haven't done before.
One area in which it may affect us directly is the resetting of counters
in the test helper, which may slow down test suite runs a bit.
* The with_files scope on the Thesis model has been fixed to correctly
deduplicate results.