Update ProQuest opt-in field based on UXWS review
Why these changes are being introduced:

UXWS has suggested changes to the ProQuest opt-in field in
consultation with Mikki, Wilder, and Katie.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-602

How this addresses that need:

This updates the wording in the label for the ProQuest opt-in field,
adjusts the styling, and adds the field to the confirmation email.

Side effects of this change:

We are now including the ThesisHelper in the ReceiptMailer, so we
can use the render_proquest_status method in the confirmation email.