Move descriptive text of ProQuest opt-in outside of field label
Why these changes are being introduced:

During a11y review, we determined that interactive content is
not advisable inside a label. Anchors and other interactive
elements can confuse screen reader navigation within the label,
and make it difficult for people to activate the corresponding
form input.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-602

How this addresses that need:

This moves the bulk of the descriptive content outside of the field
label, so the label is more direct in what it is asking of the user.
This also adds a header to the ProQuest field, which Dhruti recommended.

Side effects of this change:

None.