Add DataTables to ProQuest opt-in status report
Why these changes are being introduced:

During QA for ETD-591 and ETD-599, stakeholders noted that it would
be useful to sort on the the 'Opted in to ProQuest?' column in the
ProQuest opt-in status report.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-591
* https://mitlibraries.atlassian.net/browse/ETD-599

How this addresses that need:

This adds DataTables to the ProQuest status report, which provides
sorting on all columns, searching, and pagination. The default page
length is 'All'. This reflects the current behavior of the report,
which stakeholders have indicated meets their needs, but the
lengthMenu parameter is included in case a smaller result set is
desired.

Side effects of this change:

DataTables may be a heavier tool than we need to accommodate this
feature request. However, it shouldn't affect performance much.
The DataTables performance issues we've seen elsewhere in the app are
related to loading all data at once rather than paginating. This report
already does that, and its page load time in production is acceptable
(presumably thanks to its use of counter_cache).