Add scope note to ProQuest opt-in status report
Why these changes are being introduced:

The ProQuest opt-in status report only lists theses with advanced
degrees and files attached. These conditions are not clear in the
current UI.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-604

How this addresses that need:

This adds an alert banner clarifying the scope of the report.

Side effects of this change:

None.