Fix bug in consented_to_proquest scope ignoring null values
Why these changes are being introduced:

The consented_to_proquest scope returns thesis for which all authors
have a proquest_allowed value of true and not false. This ignores
author records with a null proquest_allowed value, so if a thesis
has multiple authors and at least one of them has not selected
an opt-in status, it will incorrectly be flagged as ready
for ProQuest export.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-588
https://mitlibraries.atlassian.net/browse/ETD-603

How this addresses that need:

This chains an additional `excluding` call to the scope that checks
for authors with null proquest_allowed, and adds regression tests
to confirm the correct behavior.

Side effects of this change:

The additional `excluding` call is a bit ugly, but after fiddling
with it for a while I decided it was more important to land a
working solution than an elegant one.