Refactor tests for consented_to_proquest scopes
Why these changes are being introduced:

Tests for the `consented_to_proquest` and `not_consented_to_proquest`
scopes alter data in fixtures used for previous assertions rather
than using fixtures that specifically meet each test case. This
can be brittle, if one were to modify the test without noticing
the changes in between assertions.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-621

How this addresses that need:

This adds more fixtures to meet each test case of the ProQuest
consent scopes.

Side effects of this change:

Updated a couple of tests to pass, including refactoring an advisor
test not to attempt to select a specific thesis attached to a
specific user. That caused the test to fail in this case, and the
new approach seems more resilient.