Generate budget report CSV for partially harvested theses
Why these changes are being introduced:

We pay ProQuest to harvest metadata from doctoral theses that have
not opted in. Each time the export job is run, we need to generate
a report of these theses for budget reconciliation.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-616

How this addresses that need:

This adds logic to generate the requisite report, which is then
generated during the ProquestExportJob and attached the
confirmation email.

Side effects of this change:

Unrelated tests in the Report model have updated due to new fixtures
added.